+++ swagger-request-validator-core/src/test/java/com/atlassian/oai/validator/OpenAPIV3RequestValidationTest.java
@@ -325,6 +325,23 @@ public class OpenAPIV3RequestValidationTest {
     }
 
     @Test
+    public void validate_withAllOfComposition_passes_whenResolveCombinatorsOptionUsed() {
+        final OpenApiInteractionValidator classUnderTest =
+                OpenApiInteractionValidator
+                        .createForSpecificationUrl("/oai/v3/api-complex-composition.yaml")
+                        .withResolveCombinators(true)
+                        .build();
+
+        final Request request = SimpleRequest.Builder
+                .post("/allOf")
+                .withContentType("application/json")
+                .withBody("{ \"stringField\": \"foo\", \"intField\": 1, \"boolField\": false }")
+                .build();
+
+        assertPass(classUnderTest.validateRequest(request));
+    }
+
+    @Test
     public void validate_withAllOfComposition_shouldFail_whenInvalidAccordingToSchema() {
         final OpenApiInteractionValidator classUnderTest =
                 OpenApiInteractionValidator
@@ -777,6 +794,22 @@ public class OpenAPIV3RequestValidationTest {
                 .build();
 
         assertFail(classUnderTest.validateRequest(request), "validation.request.body.schema.pattern");
+    }
+
+    @Test
+    public void validate_withNonRequired_minLength_shouldPass() {
+        final OpenApiInteractionValidator classUnderTest = OpenApiInteractionValidator
+                .createForSpecificationUrl("/oai/v3/api-with-minlength-properties.yaml")
+                .build();
+
+        final Request request = SimpleRequest.Builder
+                .post("/minLengthProperties")
+                .withContentType("application/json")
+                // Body missing `nonRequiredField` that also has a minLength: 1
+                .withBody("{\"requiredField\": \"foo\"}")
+                .build();
+
+        assertPass(classUnderTest.validateRequest(request));
     }
 
     private class TestValidator implements CustomRequestValidator {