+++ swagger-request-validator-springmvc/src/test/java/com/atlassian/oai/validator/springmvc/OpenApiValidationServiceTest.java
@@ -1,19 +1,18 @@
 package com.atlassian.oai.validator.springmvc;
 
 import com.atlassian.oai.validator.OpenApiInteractionValidator;
+import com.atlassian.oai.validator.example.simple.RestServiceApplication;
 import com.atlassian.oai.validator.model.Body;
+import com.atlassian.oai.validator.model.InputStreamBody;
 import com.atlassian.oai.validator.model.Request;
 import com.atlassian.oai.validator.model.Response;
 import com.atlassian.oai.validator.model.SimpleRequest;
 import com.atlassian.oai.validator.model.SimpleResponse;
 import com.atlassian.oai.validator.report.ValidationReport;
-import com.atlassian.oai.validator.example.simple.RestServiceApplication;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Iterators;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
@@ -26,7 +25,6 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.util.ReflectionUtils;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestHeader;
@@ -51,22 +49,21 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.Enumeration;
 import java.util.Map;
-import java.util.Optional;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptySet;
 import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
-import static org.hamcrest.CoreMatchers.sameInstance;
-import static org.junit.Assert.assertThat;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.when;
-import static org.springframework.test.util.ReflectionTestUtils.getField;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, properties = {"server.contextPath=/v1"})
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
+        properties = {"server.contextPath=/v1", "server.error.include-message=always"})
 public class OpenApiValidationServiceTest {
 
     @Autowired
@@ -87,20 +84,17 @@ public class OpenApiValidationServiceTest {
         return Iterators.asEnumeration(Arrays.asList(values).iterator());
     }
 
-    private static Object getInputStreamField(final Optional<Body> requestBody) {
-        return getField(getField(requestBody.get(), "content"), "in");
-    }
-
-    @Before
+    @BeforeEach
     public void setUp() {
         requestValidator = mock(OpenApiInteractionValidator.class);
         urlPathHelper = mock(UrlPathHelper.class);
         classUnderTest = new OpenApiValidationService(requestValidator, urlPathHelper);
     }
 
-    @Test(expected = NullPointerException.class)
+    @Test
     public void constructor_failsWithoutRequiredValidator() {
-        new OpenApiValidationService((OpenApiInteractionValidator) null, urlPathHelper);
+        assertThrows(NullPointerException.class,
+                () -> new OpenApiValidationService((OpenApiInteractionValidator) null, urlPathHelper));
     }
 
     @Test
@@ -132,9 +126,10 @@ public class OpenApiValidationServiceTest {
         assertThat(validationReport.hasErrors(), is(false));
     }
 
-    @Test(expected = NullPointerException.class)
-    public void buildRequest_failsWithoutRequiredRequest() throws IOException {
-        classUnderTest.buildRequest(null);
+    @Test
+    public void buildRequest_failsWithoutRequiredRequest() {
+        assertThrows(NullPointerException.class,
+                () -> classUnderTest.buildRequest(null));
     }
 
     @Test
@@ -143,7 +138,8 @@ public class OpenApiValidationServiceTest {
         when(servletRequest.getMethod()).thenReturn("GET");
         when(servletRequest.getQueryString()).thenReturn("");
         when(urlPathHelper.getPathWithinApplication(servletRequest)).thenReturn("/swagger-request-validator");
-        final ServletInputStream inputStream = mock(ServletInputStream.class);
+        final String content = "hello";
+        final ServletInputStream inputStream = new ServletInputStreamMock(content.getBytes(StandardCharsets.UTF_8));
         when(servletRequest.getInputStream()).thenReturn(inputStream);
         when(servletRequest.getParameterNames()).thenReturn(asEnumeration("not-a-query-parameter"));
         when(servletRequest.getHeaderNames()).thenReturn(Collections.emptyEnumeration());
@@ -152,7 +148,9 @@ public class OpenApiValidationServiceTest {
 
         assertThat(result.getPath(), equalTo("/swagger-request-validator"));
         assertThat(result.getMethod(), equalTo(Request.Method.GET));
-        assertThat(getInputStreamField(result.getRequestBody()), sameInstance(inputStream));
+        final Body body = result.getRequestBody().get();
+        assertThat(body, instanceOf(InputStreamBody.class));
+        assertThat(body.toString(StandardCharsets.UTF_8), equalTo(content));
         assertThat(result.getHeaders().size(), equalTo(0));
         assertThat(result.getQueryParameters().size(), equalTo(0));
     }
@@ -164,7 +162,8 @@ public class OpenApiValidationServiceTest {
         when(servletRequest.getQueryString())
                 .thenReturn("query1=QUERY_ONE&&query2=query_two&query2=QUERY_TWO&");
         when(urlPathHelper.getPathWithinApplication(servletRequest)).thenReturn("/swagger-request-validator");
-        final ServletInputStream inputStream = mock(ServletInputStream.class);
+        final String content = "hello";
+        final ServletInputStream inputStream = new ServletInputStreamMock(content.getBytes(StandardCharsets.UTF_8));
         when(servletRequest.getInputStream()).thenReturn(inputStream);
         when(servletRequest.getParameterNames())
                 .thenReturn(asEnumeration("query1", "query2", "query3"));
@@ -183,7 +182,9 @@ public class OpenApiValidationServiceTest {
 
         assertThat(result.getPath(), equalTo("/swagger-request-validator"));
         assertThat(result.getMethod(), equalTo(Request.Method.POST));
-        assertThat(getInputStreamField(result.getRequestBody()), sameInstance(inputStream));
+        final Body body = result.getRequestBody().get();
+        assertThat(body, instanceOf(InputStreamBody.class));
+        assertThat(body.toString(StandardCharsets.UTF_8), equalTo(content));
         assertThat(result.getHeaders().size(), equalTo(2));
         assertThat(result.getHeaderValues("header1"),
                 equalTo(asList("HEADER_ONE")));
@@ -220,10 +221,11 @@ public class OpenApiValidationServiceTest {
         assertThat(result.getQueryParameters().size(), equalTo(0));
     }
 
-    @Test(expected = NullPointerException.class)
-    public void buildResponse_failsWithoutRequiredResponse() throws IOException {
+    @Test
+    public void buildResponse_failsWithoutRequiredResponse() {
         // expect:
-        classUnderTest.buildResponse(null);
+        assertThrows(NullPointerException.class,
+                () -> classUnderTest.buildResponse(null));
     }
 
     @Test
@@ -322,15 +324,15 @@ public class OpenApiValidationServiceTest {
 
         // then: assert the values that spring has been set in the controller method
         final Map springRequest = (Map) responseEntity.getBody().get("springRequest");
-        Assert.assertThat(springRequest.get("pathVariable"), is("path variable"));
-        Assert.assertThat(springRequest.get("queryParam"), is("query param"));
-        Assert.assertThat(springRequest.get("headerValue"), is("header value"));
+        assertThat(springRequest.get("pathVariable"), is("path variable"));
+        assertThat(springRequest.get("queryParam"), is("query param"));
+        assertThat(springRequest.get("headerValue"), is("header value"));
 
         // and: assert the values that the validation service has been set from the servlet request
         final Map validationRequest = (Map) responseEntity.getBody().get("validationRequest");
-        Assert.assertThat(validationRequest.get("path"), is("/test controller/path variable"));
-        Assert.assertThat(validationRequest.get("queryParam"), is("query param"));
-        Assert.assertThat(validationRequest.get("headerValue"), is("header value"));
+        assertThat(validationRequest.get("path"), is("/test controller/path variable"));
+        assertThat(validationRequest.get("queryParam"), is("query param"));
+        assertThat(validationRequest.get("headerValue"), is("header value"));
     }
 
     @Test